Gem rename POSIX::Spawn::Child -> Progeny::Command
Move name away from posix spawn since the gem no longer has anything to
do with the `posix_spawn` system call
