--- lib/posix-spawn.rb
@@ -1,1 +1,0 @@
-require "posix/spawn"
