--- lib/posix/spawn.rb
@@ -1,2 +1,0 @@
-require 'posix/spawn/version'
-require 'posix/spawn/child'
