--- lib/posix/spawn/version.rb
@@ -1,5 +1,0 @@
-module POSIX
-  module Spawn
-    VERSION = '0.3.15'
-  end
-end
