+++ lib/progeny.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative "progeny/version"
+require_relative "progeny/command"
