+++ lib/progeny/version.rb
@@ -1,0 +1,3 @@
+module Progeny
+  VERSION = '0.3.15'
+end
