+++ test/test_command.rb
@@ -1,9 +1,8 @@
-# coding: UTF-8
+# frozen_string_literal: true
+
 require 'test_helper'
 
-class ChildTest < Minitest::Test
-  include POSIX::Spawn
-
+class CommandTest < Minitest::Test
   # Become a new process group.
   def setup
     Process.setpgrp
@@ -33,106 +32,102 @@ class ChildTest < Minitest::Test
   # be a IO.popen under Ruby >= 1.9 since it also supports :pgroup.
   def assert_process_group_reaped(pgid)
     command = "ps axo pgid,pid,args | grep '^#{pgid} ' | grep -v '^#{pgid} #$$'"
-    procs = POSIX::Spawn::Child.new(command, :pgroup => true).out
+    procs = Progeny::Command.new(command, :pgroup => true).out
     assert procs.empty?, "Processes in group #{pgid} still running:\n#{procs}"
   end
 
-  def test_sanity
-    assert_same POSIX::Spawn::Child, Child
-  end
-
   def test_argv_array_execs
-    p = Child.new('printf', '%s %s %s', '1', '2', '3 4')
+    p = Progeny::Command.new('printf', '%s %s %s', '1', '2', '3 4')
     assert p.success?
     assert_equal "1 2 3 4", p.out
   end
 
   def test_argv_string_uses_sh
-    p = Child.new("echo via /bin/sh")
+    p = Progeny::Command.new("echo via /bin/sh")
     assert p.success?
     assert_equal "via /bin/sh\n", p.out
   end
 
   def test_stdout
-    p = Child.new('echo', 'boom')
+    p = Progeny::Command.new('echo', 'boom')
     assert_equal "boom\n", p.out
     assert_equal "", p.err
   end
 
   def test_stderr
-    p = Child.new('echo boom 1>&2')
+    p = Progeny::Command.new('echo boom 1>&2')
     assert_equal "", p.out
     assert_equal "boom\n", p.err
   end
 
   def test_status
-    p = Child.new('exit 3')
+    p = Progeny::Command.new('exit 3')
     assert !p.status.success?
     assert_equal 3, p.status.exitstatus
   end
 
   def test_env
-    p = Child.new({ 'FOO' => 'BOOYAH' }, 'echo $FOO')
+    p = Progeny::Command.new({ 'FOO' => 'BOOYAH' }, 'echo $FOO')
     assert_equal "BOOYAH\n", p.out
   end
 
   def test_chdir
-    p = Child.new("pwd", :chdir => File.dirname(Dir.pwd))
+    p = Progeny::Command.new("pwd", :chdir => File.dirname(Dir.pwd))
     assert_equal File.dirname(Dir.pwd) + "\n", p.out
   end
 
   def test_input
     input = "HEY NOW\n" * 100_000 # 800K
-    p = Child.new('wc', '-l', :input => input)
+    p = Progeny::Command.new('wc', '-l', :input => input)
     assert_equal 100_000, p.out.strip.to_i
   end
 
   def test_max
-    child = Child.build('yes', :max => 100_000)
-    assert_raises(MaximumOutputExceeded) { child.exec! }
+    child = Progeny::Command.build('yes', :max => 100_000)
+    assert_raises(Progeny::MaximumOutputExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped Process.pid
   end
 
   def test_max_pgroup_kill
-    child = Child.build('yes', :max => 100_000, :pgroup_kill => true)
-    assert_raises(MaximumOutputExceeded) { child.exec! }
+    child = Progeny::Command.build('yes', :max => 100_000, :pgroup_kill => true)
+    assert_raises(Progeny::MaximumOutputExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped child.pid
   end
 
   def test_max_with_child_hierarchy
-    child = Child.build('/bin/sh', '-c', 'true && yes', :max => 100_000)
-    assert_raises(MaximumOutputExceeded) { child.exec! }
+    child = Progeny::Command.build('/bin/sh', '-c', 'true && yes', :max => 100_000)
+    assert_raises(Progeny::MaximumOutputExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped Process.pid
   end
 
   def test_max_with_child_hierarchy_pgroup_kill
-    child = Child.build('/bin/sh', '-c', 'true && yes', :max => 100_000, :pgroup_kill => true)
-    assert_raises(MaximumOutputExceeded) { child.exec! }
+    child = Progeny::Command.build('/bin/sh', '-c', 'true && yes', :max => 100_000, :pgroup_kill => true)
+    assert_raises(Progeny::MaximumOutputExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped child.pid
   end
 
   def test_max_with_stubborn_child
-    child = Child.build("trap '' TERM; yes", :max => 100_000)
-    assert_raises(MaximumOutputExceeded) { child.exec! }
+    child = Progeny::Command.build("trap '' TERM; yes", :max => 100_000)
+    assert_raises(Progeny::MaximumOutputExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped Process.pid
   end
 
   def test_max_with_stubborn_child_pgroup_kill
-    child = Child.build("trap '' TERM; yes", :max => 100_000, :pgroup_kill => true)
-    assert_raises(MaximumOutputExceeded) { child.exec! }
+    child = Progeny::Command.build("trap '' TERM; yes", :max => 100_000, :pgroup_kill => true)
+    assert_raises(Progeny::MaximumOutputExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped child.pid
   end
 
   def test_max_with_partial_output
-    p = Child.build('yes', :max => 100_000)
+    p = Progeny::Command.build('yes', :max => 100_000)
     assert_nil p.out
-    assert_raises MaximumOutputExceeded do
+    assert_raises Progeny::MaximumOutputExceeded do
       p.exec!
     end
     assert_output_exceeds_repeated_string("y\n", 100_000, p.out)
@@ -141,8 +136,8 @@ class ChildTest < Minitest::Test
   end
 
   def test_max_with_partial_output_long_lines
-    p = Child.build('yes', "nice to meet you", :max => 10_000)
-    assert_raises MaximumOutputExceeded do
+    p = Progeny::Command.build('yes', "nice to meet you", :max => 10_000)
+    assert_raises Progeny::MaximumOutputExceeded do
       p.exec!
     end
     assert_output_exceeds_repeated_string("nice to meet you\n", 10_000, p.out)
@@ -152,8 +147,8 @@ class ChildTest < Minitest::Test
 
   def test_timeout
     start = Time.now
-    child = Child.build('sleep', '1', :timeout => 0.05)
-    assert_raises(TimeoutExceeded) { child.exec! }
+    child = Progeny::Command.build('sleep', '1', :timeout => 0.05)
+    assert_raises(Progeny::TimeoutExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped Process.pid
     assert (Time.now-start) <= 0.2
@@ -161,30 +156,30 @@ class ChildTest < Minitest::Test
 
   def test_timeout_pgroup_kill
     start = Time.now
-    child = Child.build('sleep', '1', :timeout => 0.05, :pgroup_kill => true)
-    assert_raises(TimeoutExceeded) { child.exec! }
+    child = Progeny::Command.build('sleep', '1', :timeout => 0.05, :pgroup_kill => true)
+    assert_raises(Progeny::TimeoutExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped child.pid
     assert (Time.now-start) <= 0.2
   end
 
   def test_timeout_with_child_hierarchy
-    child = Child.build('/bin/sh', '-c', 'true && sleep 1', :timeout => 0.05)
-    assert_raises(TimeoutExceeded) { child.exec! }
+    child = Progeny::Command.build('/bin/sh', '-c', 'true && sleep 1', :timeout => 0.05)
+    assert_raises(Progeny::TimeoutExceeded) { child.exec! }
     assert_process_reaped child.pid
   end
 
   def test_timeout_with_child_hierarchy_pgroup_kill
-    child = Child.build('/bin/sh', '-c', 'true && sleep 1', :timeout => 0.05, :pgroup_kill => true)
-    assert_raises(TimeoutExceeded) { child.exec! }
+    child = Progeny::Command.build('/bin/sh', '-c', 'true && sleep 1', :timeout => 0.05, :pgroup_kill => true)
+    assert_raises(Progeny::TimeoutExceeded) { child.exec! }
     assert_process_reaped child.pid
     assert_process_group_reaped child.pid
   end
 
   def test_timeout_with_partial_output
     start = Time.now
-    p = Child.build('echo Hello; sleep 1', :timeout => 0.05, :pgroup_kill => true)
-    assert_raises(TimeoutExceeded) { p.exec! }
+    p = Progeny::Command.build('echo Hello; sleep 1', :timeout => 0.05, :pgroup_kill => true)
+    assert_raises(Progeny::TimeoutExceeded) { p.exec! }
     assert_process_reaped p.pid
     assert_process_group_reaped Process.pid
     assert (Time.now-start) <= 0.2
@@ -200,7 +195,7 @@ class ChildTest < Minitest::Test
         echo stuff on stderr 1>&2;
       done
     "
-    p = Child.new(command, :input => input)
+    p = Progeny::Command.new(command, :input => input)
     assert_equal input.size, p.out.size
     assert_equal input.size, p.err.size
     assert p.success?
@@ -208,19 +203,19 @@ class ChildTest < Minitest::Test
 
   def test_input_cannot_be_written_due_to_broken_pipe
     input = "1" * 100_000
-    p = Child.new('false', :input => input)
+    p = Progeny::Command.new('false', :input => input)
     assert !p.success?
   end
 
   def test_utf8_input
     input = "hålø"
-    p = Child.new('cat', :input => input)
+    p = Progeny::Command.new('cat', :input => input)
     assert p.success?
   end
 
   def test_utf8_input_long
     input = "hålø" * 10_000
-    p = Child.new('cat', :input => input)
+    p = Progeny::Command.new('cat', :input => input)
     assert p.success?
   end
