+++ test/test_helper.rb
@@ -1,9 +1,7 @@
-require 'minitest/autorun'
-require 'posix-spawn'
+# frozen_string_literal: true
 
-if Minitest.const_defined?('Test')
-  # We're on Minitest 5+. Nothing to do here.
-else
-  # Minitest 4 doesn't have Minitest::Test yet.
-  Minitest::Test = MiniTest::Unit::TestCase
-end
+$LOAD_PATH.unshift File.expand_path("../lib", __dir__)
+require "progeny"
+
+require "minitest/autorun"
+
